# backend/docs/swagger.yaml
basePath: /
definitions:
...
  discovery.Item:
    properties:
      city:
        type: string
      detail:
        type: string
      icon:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      lat:
        type: number
      lng:
        type: number
      meta:
        type: string
      name:
        type: string
      type:
        type: string
      validEndDate:
        type: string
      validStartDate:
        type: string
    type: object
  httpx.Error:
    properties:
      detail:
        type: string
    type: object
  marketplace.Application:
    properties:
      id:
        type: string
      jobId:
        type: string
      message:
        type: string
      price:
        type: integer
      provider:
        $ref: '#/definitions/marketplace.Provider'
    type: object
  marketplace.Category:
    properties:
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      sub:
        type: string
    type: object
  marketplace.Job:
    properties:
      applicants:
        type: integer
      area:
        type: string
      budget:
        type: integer
      category:
        type: string
      description:
        type: string
      id:
        type: string
      note:
        type: string
      provider:
        description: provider.name, null when unassigned
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  marketplace.JobInput:
    properties:
      area:
        type: string
      budget:
        type: integer
      category:
        type: string
      description:
        type: string
      title:
        type: string
    type: object
  marketplace.Provider:
    properties:
      about:
        type: string
      city:
        type: string
      color:
        type: string
      distanceKm:
        type: integer
      from:
        type: integer
      id:
        description: provider slug, used in /providers/[id] routes
        type: string
      initials:
        type: string
      jobs:
        type: integer
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      rating:
        type: number
      reviews:
        items:
          $ref: '#/definitions/marketplace.Review'
        type: array
      trade:
        type: string
      verified:
        type: boolean
      years:
        type: integer
    type: object
  marketplace.Review:
    properties:
      name:
        type: string
      stars:
        type: integer
      text:
        type: string
    type: object
  marketplace.Thread:
    properties:
      color:
        type: string
      id:
        type: string
      initials:
        type: string
      messages:
        items:
          $ref: '#/definitions/marketplace.ThreadMessage'
        type: array
      name:
        type: string
      preview:
        type: string
      time:
        type: string
      unread:
        type: integer
    type: object
  marketplace.ThreadMessage:
    properties:
      me:
        type: boolean
      text:
        type: string
      time:
        type: string
    type: object
  marketplace.Transaction:
    properties:
      amount:
        type: integer
      date:
        type: string
      id:
        type: string
      label:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: LHCHub home-services marketplace API (Go backend). Clients post jobs,
    vetted providers apply, and payment is held in escrow.
  termsOfService: http://swagger.io/terms/
  title: LHCHub API
  version: "1.0"
paths:
...
  /api/discovery/items/:
    get:
      description: Returns Local Discovery POIs filtered by city and type.
      parameters:
      - description: City slug to filter by
        in: query
        name: city
        type: string
      - description: 'POI type: things_to_do|events|promos|news'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/discovery.Item'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List discovery items
      tags:
      - discovery
  /api/marketplace/applications/:
    get:
      description: Returns job applications, optionally filtered by job id.
      parameters:
      - description: Job id filter
        in: query
        name: jobId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace.Application'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List applications
      tags:
      - marketplace
  /api/marketplace/categories/:
    get:
      description: Returns all marketplace job categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List categories
      tags:
      - marketplace
  /api/marketplace/jobs/:
    get:
      description: Returns jobs, optionally filtered by status and category.
      parameters:
      - description: Job status filter
        in: query
        name: status
        type: string
      - description: Category filter
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace.Job'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List jobs
      tags:
      - marketplace
    post:
      consumes:
      - application/json
      description: Posts a new job. Requires authentication.
      parameters:
      - description: Job to create
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/marketplace.JobInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/marketplace.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Create a job
      tags:
      - marketplace
  /api/marketplace/jobs/{id}/:
    get:
      description: Returns a single job by id.
      parameters:
      - description: Job id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/marketplace.Job'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Get a job
      tags:
      - marketplace
  /api/marketplace/providers/:
    get:
      description: Returns providers, optionally filtered by city, trade, verified
        status, and max price.
      parameters:
      - description: City filter
        in: query
        name: city
        type: string
      - description: Trade/category filter
        in: query
        name: trade
        type: string
      - description: Set to true to show only verified providers
        in: query
        name: verified
        type: string
      - description: Maximum price filter
        in: query
        name: maxPrice
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace.Provider'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List providers
      tags:
      - marketplace
  /api/marketplace/providers/{slug}/:
    get:
      description: Returns a single provider by slug.
      parameters:
      - description: Provider slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/marketplace.Provider'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Get a provider
      tags:
      - marketplace
  /api/marketplace/threads/:
    get:
      description: Returns the user's message threads.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace.Thread'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List message threads
      tags:
      - marketplace
  /api/marketplace/transactions/:
    get:
      description: Returns the user's wallet transactions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List wallet transactions
      tags:
      - marketplace
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"